getting rid of shared module to make github actions happy
